Change packSize type from Double to Integer in PurchaseOrder
Updated the packSize field in PurchaseOrder entity from Double to Integer for type consistency. Adjusted related test cases in controller, repository, and service tests to use Integer values for packSize.

+++ lis/src/main/java/com/biotech/lis/Entity/PurchaseOrder.java
@@ -20,7 +20,7 @@ public class PurchaseOrder {
     
     private Integer brandId;
     
-    private Double packSize;
+    private Integer packSize;
     
     private Integer quantity;
